Move LowLevelDeserializationError to scylla-cql-core
Ultimately we want to move whole `frame_errors` module, but we can't do
that yet because of dependencies on serialization / deserialization.
We can move `LowLevelDeserializationError` because it only depends on
stuff that was already moved. We have to move it, because without it we
can't move stuff (like RawValue) needed by serialization /
deserialization.
