+++ scylla-cql-core/src/frame/mod.rs
@@ -1,3 +1,4 @@
+pub mod frame_errors;
 pub mod response;
 pub mod types;
